HMR development stats: include `updatedModules` for App Router and Turbopack changes (#59785)
This includes the list of changed modules in the `client-hmr-latency`
event for changes in app router pages as well as when using Turbopack.

For Turbopack, this list is derived from the data structure included in
the `turbopack-message` event. Each of these is originally a module id,
which includes additional context information. For analytics purposes
this is removed (and could maybe surfaced in the future once the webpack
data is also normalized).


Closes PACK-2167