Implements a concept and set of APIs related to "Dynamic Rendering". Previously this was talked about externally in Docs but implemented as a staticGenerationBailout. There are however different semantics we want to have for dynamic using htings like `unstable_noStore` vs `headers()` for instance and having APIs oriented around these semantics will make future planned changes easier to implement correctly.
The two main semantics are "the currently executing scope should be dynamic" and "some dynamic data source is being read"

The former will eventually be a noop if we make dynamic the default. In accordance with this the dynamic intent is ignored inside a cache scope from `unstable_cache`. This is a very weak form of dynamic scoping and makes sense because it does not entail a request specific data read.

The latter semantic is a stronger form. when doing any kind of static generation we disallow reading dynamic data sources such as headers, cookies, and searchParams. If these are read during a static render the render will bailout or postpone (depending on whether we are using PPR or not)

force-static should still forcibly prevent dynamic data reads. This is now implementation outside of the triggering of a dynamic data read.

This PR does not actually change yet how we define whether the result of a render was fully, partially, or not static. This will come in a follow up change where we can more faithfully track whether dynamic APIs were used separately from whether the render postpones.