Tracing: attach Turbopack session value to root span
This allows us to more accurately attribute whether a trace was created in a Turbopack session by doing so at trace creation time rather than trace upload time.

Test Plan: Configured next to send traces to a local server and verified this information is attached to the top-level of the trace.