turbo LSP + VSCode extension (vercel/turbo#6543)
### Description

This is now ready for review. I have included a vsix file below that you
can download and test, simply add it to VSCode. You will need a recent
version (ie. what is currently in main) to use it properly so I suggest
pulling out a recent master, building a binary, and copying it to
`node_modules/turbo-<platform>` for whatever project you wish to test it
with.


[turbo-vsc-1.0.0.vsix.zip](https://github.com/vercel/turbo/files/13887415/turbo-vsc-1.0.0.vsix.zip)

Closes TURBO-1737

---------

Co-authored-by: Alexander Lyon <Alexander Lyon>