fix a hanging issue with strongly consistent reads (vercel/turbo#7067)
### Description

This fixes a bug where inactive tasks get stuck in "in progress" which
causes the updateInfo stream to be incorrect (not ending).

It can happen when
* a task reads something strongly consistent
* the task (subgraph) is not active (e.g. was removed due to changes)

In this case the strongly consistent read will never finish as it waits
for the subgraph to finish, but dirty tasks in that subgraph won't be
scheduled as the subgraph is not active.

This PR fixes that by making a subgraph that is currently strongly
consistently read temporarily active (while being read).

That's similar to normal read scheduling dirty tasks.

We never want in progress tasks to be blocked by anything. Any blocker
should be resolved by executing tasks.

Potential alternative: We could also change dirty tasks to not count as
pending. That would also allow any strongly consistent read to complete,
but it won't be consistent with changes that caused these tasks to be
dirty.


Closes PACK-2268