Fix sitemap generateSitemaps support for string id (#61088)
Fixes the string id that broken when sitemap is optimized to static
route.

When sitemap is optimized to static route in production, the route
argument is changed from `[[...__metadata_id__]]` to
`[__metadata_id__]`, so the type of it is also changed from array to
string that should reflect in the loader code.

Fixes #60894
Closes NEXT-2154